Broke up type cycle and made it readonly
