--- packages/typescript-estree/src/log.ts
@@ -1,3 +1,0 @@
-import debug from 'debug';
-
-export const log = debug('typescript-eslint:typescript-estree:parser');
